Still improving the reliability of the checkConstraintIntegrity method. some changes occured :
A new test class, TestPoint, is added, to check that the methods in MyPoint are reliable, particulaarly the comparison and equality checks.

Minor changes occured in MyEdge, related to the intersection computation. The related methods are more documented now.

There were a bug in ConstrainedMesh.addEdgeToLeftSortedList

ConstrainedDelaunay is still evolving, as i'm  adding test cases on and on. I'm currently stuck with the last one : a bug is still present while using the forceConstrainedMeshIntegrity... and just appears when using random edges... I still need to add particular cases, so, to understand the bug more clearly...