Added some tests on edges, triangles and bounding boxes.
A method in MyMesh was definitely not safe, as it used an external int
to control a loop on an array, and not the length of this given array...

about to complete the first compile-safe integration of triangulation in
the constrained mesh. It's not completely ok, as we need to add
verifications to avoid inserting edges that would intersect with the
constraint edges. codes are commented in the commit, in order to permit
the users to pass the compilation.... ;-)