Following the arrival of the interface Comparable in Point and DelaunayTriangle, we introduce it in Edge. This implementation is arguable, and not trivial, as many sort methods are used on Edges. I've decided to implement it in relation with the left-right sorting method.
Consequently, many changes appear in ConstrainedMesh. We are handling
many lists that are sorted with objects that directly use the interface
Comparable. Consequently, we can directly use the
Collections.sort(List<? extends Comparable>) and
Collections.binarySearch(List<T extends Comparable>, T). This way, we
can use genericity and have unique (and simpler) methods for some of
these operations, relying on the Java API (which will offer the same
performances in the worst case...).

Tests evolve consequently.