Add a method to check if there is an intersection between an edge and the edges which are upper and lowerthan a point in a vertical list.
the design of vertical list has changed. Indeed, we'll need to perform these comparisons many times with the same point and list during the mesh computation. Consequently, I've decided to add volatile attributes in VerticalList. They are used to remember which point has been used, and whatare the upper and lower edge in this list for this point. They are removed when a hard operation is made on the list (new sort, insertion, deletion, modification of the comparator).

Tests added for these methods.

A new error code has been added to DelaunayError. It is used in the constrained mesh computation, whil forcing the integrity of the constraints.

The implementation of the constrained mesh genration algorithm has begun. More to come, it's not ready yet. The new class TestConstrainedMesh will gather the tests for the mesh computation. There is only one for now, and it's still broken...