Many changes occur in this commit...
a further study of the previous triangulation algorithm show that when using it, we could forget some of the triangles of our mesh. We c'ant accept that. Consequently, we have included some modifications in the algorithm. It is more complex now, but works well in many more cases...

Changelog : 

DelaunayTriangle :
the getPoint method has been fixed, it made comparisons with == rather than equals on Point instances. 
Some utility methods enter in the API (getLastEdge, getOppositeEdge and getLeftMost)
The compareTo method has been fixed, as it were likely to return 0 with non-overlapping area...

Unit tests are included to check these new events and cases.

Edge : 

an attribute "degenerated" is added. The current constrained triangulation algorithm need such an attribute to make the distinction between edges linked to the mesh and at leas one triangle, and edges that are just linked to the mesh (that are consequently considered as "degenerated").
A utility method appears (getOtherTriangle).

Added matching unit test.

ConstrainedMesh : 

The diff is really huge, here... The public API is not really modified, but some private methods appear.

Added a method to retrieve an edge from the constraints or a list of edge, to use an existing edge rather than a newly created one.
Added a separate method to generate the first triangle (or the two first edges whee such a triangle can't be created because of existing constraints)
The codes that insert a point into the mesh (insertPoint) has been deeply modified. We are now able to treat the cases where a point cant be directly linked to the mesh with a triangle, but just with an edge (which is marked as degenerated). Unit tests have been included to stress as many cases as possible. We'll certainly need more work to find some special cases, in order to be sure that the implementation is solid.