Some debugging changes in the implementations
ConstrainedMesh
====================

 * Adding control structures when manipulating new edges and boundaryEdges, to be sur we don't have duplicated. Not very clean, currently, it will need further investigation.
 * The first triangle was not correctly processed ! We didn't manage the case where p1p3 was intersecting a constraint from p2. Fixed.

TestConstrainedMesh
====================

I'm currently working with the datas that come from the chezine. As i'm facing bugs with them, I try to reproduce them in my unit tests.

 * test3HorizontalPoints and testCross perform both new tests were a special configuration is managed. There are here for safety first.
 * testTwoNeighbourConstraints manages a special case, where the first triangle can't be built.

TestVerticalList
====================

Contains a new test, dedicated to the removal of an edge from its right point in the list.

Edge
====================

ThegtPointFromitsX method has been modified. Before this changeset, the point returned for a verticaledge was the left one. It's now the right one, as it is the one needed in the verticalList.