The vertical comparision is not used with the same goals during the intersection computation and during the triangulation. Consquently, the preivous fix on the verticalSort is invalid, as it breaks the intersection algorithm.
A work around, currently, is to compute the intersection between the boundary edges (all of them :-( ) and the edge to be added to the mesh. It's really long, and performances are heavily affected.

We must notice that this workaround fixes some strange behaviours that occured during the triangulation. We still encounter some problems with the first points of the mesh, though... More fix to come !