The optimization phase has begun.
TestConstrainedMesh :
Adding a test from the ones performed on MyMesh, to be sure we ensure the same features...

ConstrainedMesh :
the triangles are not sorted anymore. The inner edges will not either, but this modification forces us to perform deeper changes.

TestTriangle :
As the triangles are not sorted anymore, a test was just begun useless.