swapTriangle() has been split in two methods, as the flip flap operation could (and will) be used independently of the delaunay property test.
I'm clearly aware that the two methods should be migrated into DEdge, as they are clearly operations done on edges, not on the mesh. However, there is a side effect i'm not able to understand, and migrating the codes break test. I let them here for now, and will try again later to make the thinds cleaner.