Add a method in constrained mesh that reverts the flip flap operation. The use of a flipFlap and then of  a revertFlipFlap will ensure that the coherence of the data is kept. It means that geometries are still associated to the same GID and to the same reference in memory. It still have a (minor) border effect, as the orientation of the edge where the flip-flaps have been processed is changed by this coupled operation.
Add a test to check that the behaviour of this method is the one we want.

We are going to use maps instead of lists for handling triangles during the refinement. Our goal is to simplify the updates in the structures. We introduce three new attributes in ConstrainedMesh, and some methods to handle them. refineMesh is modified consequently. 

More changes to come, as we not ensure yet the correct use of these structures durinf the refinement.